+++ src/libslic3r/Preset.cpp
@@ -902,7 +902,7 @@ static std::vector<std::string> s_Preset_print_options {
     "wave_overhang_floor_use_hilbert", "wave_overhang_floor_hilbert_layers",
     "wave_overhang_floor_hilbert_density",
     "wave_overhang_floor_print_speed", "wave_overhang_floor_perimeter_speed", "wave_overhang_floor_fan_speed",
-    "wave_overhang_floor_aux_fan_speed",
+    "wave_overhang_floor_aux_fan_speed", "wave_overhang_floor_speed_ramp",
     "wave_overhang_min_angle",
     "wave_overhang_spacing_mode", "wave_overhang_seam_mode",
     "wave_overhang_debug_gcode", "wave_overhang_min_length",